+++ contracts/testing/TestGovernance.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.3;
 import "../Governance.sol";
 
 contract TestGovernance is Governance {
-    event Received(address, uint);
+    event Received(address, uint256);
 
     receive() external payable {
         emit Received(msg.sender, msg.value);
